Add files via upload
To ensure transparency and replicability of the results in this study, the R code used is described and provided here. Due to data privacy purposes, the dataset utilized is unavailable to the public, limiting full reproducibility to those with access to the data. Descriptions of the provided files are outlined below for clarity and ease of use. 
File 1 – Data preparation. This code cleaned and prepared the data for use in estimating the random forest model. Steps included creating outcome variables, managing missing data, and removing variables from the dataset that were not used in the initial screening process. Additionally, individuals who did not meet the specified criteria were excluded from the dataset; however, the code related to this exclusion process has been removed from the provided syntax due to confidentiality reasons.
File 2 – Random forest model. This file contains the implementation of the random forest model with cforest from the party package (version 1.3-14.; Hothorn et al., 2023), including 10-fold cross-validation. The model was tuned with 500 trees. The output "Yhat" is a vector storing predicted values for the response variable FULLFORT_RS (completed military training) across each fold in the cross-validation. These predictions can be used to calculate the raw confusion matrix of the model, although these results were not included in the paper. The "VARIMP" matrix stores the variable importance metrics for each variable in the model across each fold in the cross-validation. Each row corresponds to a variable, and each column corresponds to a fold. This data was later utilized to visualize variable importance in Figure 2. The "MODEL" object contains the trained Random Forest model for each fold in the cross-validation. This object was later used to making new predictions with different probability thresholds.
File 3 – Raw confusion matrix. This file contains code to make a confusion matrix with the raw format of the final predictions generated by the random forest model in “Yhat”. These predictions are aggregated using a weighted average across the ensemble of trees (Hothorn et al., 2004). Although not explicitly referenced in the study, this file provides insights into the initial model predictions before tuning probability thresholds. The study primarily focuses on the predictions made with tuned probability thresholds. 
File 4 – Adjusted probability thresholds. This file utilizes “MODEL” with the stored trained Random Forest to generate tables containing confusion matrices at different probability thresholds (.3, .4, .5, and .6). 
File 5 – Variable importance plot. This code utilizes the "VARIMP" matrix to create a data frame, which is then used to generate Figure 2—a boxplot visualizing the importance of each variable in the model. The boxplots show the range of importance values across the 10 folds.
File 6 – Pruned tree with the best predictors. This code uses a subset of the most important predictors to visualize their roles and interactions in predicting the outcome. It utilizes the ctree function from the party package (version 1.3-14.; Hothorn et al., 2023)  to make a conditional inference tree.